requirements.txt
discord.py
sympy
matplotlib
flask